Fix web build
