+++ src/logger/bitdriftTransport.ts
@@ -1,0 +1,24 @@
+import {
+  debug as bdDebug,
+  error as bdError,
+  info as bdInfo,
+  warn as bdWarn,
+} from '@bitdrift/react-native'
+
+import {LogLevel, Transport} from './index'
+
+export function createBitdriftTransport(): Transport {
+  return (level, message) => {
+    const log = (
+      {
+        [LogLevel.Debug]: bdDebug,
+        [LogLevel.Info]: bdInfo,
+        [LogLevel.Log]: bdInfo,
+        [LogLevel.Warn]: bdWarn,
+        [LogLevel.Error]: bdError,
+      } as const
+    )[level]
+
+    log(message.toString())
+  }
+}
