+++ src/logger/bitdriftTransport.web.ts
@@ -1,0 +1,7 @@
+import {Transport} from './index'
+
+export function createBitdriftTransport(): Transport {
+  return (_level, _message) => {
+    // noop
+  }
+}
