+++ src/logger/index.ts
@@ -1,9 +1,3 @@
-import {
-  debug as bdDebug,
-  error as bdError,
-  info as bdInfo,
-  warn as bdWarn,
-} from '@bitdrift/react-native'
 import format from 'date-fns/format'
 import {nanoid} from 'nanoid/non-secure'
 
@@ -12,6 +6,7 @@ import {DebugContext} from '#/logger/debugContext'
 import {add} from '#/logger/logDump'
 import {Sentry} from '#/logger/sentry'
 import * as env from '#/env'
+import {createBitdriftTransport} from './bitdriftTransport'
 
 export enum LogLevel {
   Debug = 'debug',
@@ -21,7 +16,7 @@ export enum LogLevel {
   Error = 'error',
 }
 
-type Transport = (
+export type Transport = (
   level: LogLevel,
   message: string | Error,
   metadata: Metadata,
@@ -135,20 +130,6 @@ export const consoleTransport: Transport = (
   } else {
     log(`${format(timestamp, 'HH:mm:ss')} ${message.toString()}${extra}`)
   }
-}
-
-export const bitdriftTransport: Transport = (level, message) => {
-  const log = (
-    {
-      [LogLevel.Debug]: bdDebug,
-      [LogLevel.Info]: bdInfo,
-      [LogLevel.Log]: bdInfo,
-      [LogLevel.Warn]: bdWarn,
-      [LogLevel.Error]: bdError,
-    } as const
-  )[level]
-
-  log(message.toString())
 }
 
 export const sentryTransport: Transport = (
@@ -349,7 +330,7 @@ export class Logger {
 export const logger = new Logger()
 
 if (!env.IS_TEST) {
-  logger.addTransport(bitdriftTransport)
+  logger.addTransport(createBitdriftTransport())
 }
 
 if (env.IS_DEV && !env.IS_TEST) {
